Upgrade to v6.0.3 and edit the Dockerfile (#6)
Been working on Docker for a few days. So far I've successfully deployed
the Dockerfile here to coherence and fly.

- It doesn't seem like we need `python3` or `build-essential` anymore.
Things have been working without them
- I can't get `apt-get clean` to prove that it saves any space. Maybe we
can pair on this again with dive to be sure @Josh-Walker-GM
- Similarly, I can't get `yarn cache clean` to prove that is saves any
space
- In some places, we had `COPY <src> .` but in others we would write out
the full name of the file. I've tried to make them consistent
- The build stages can use the CLI but the prod stage can't. I changed
the build stages to `yarn rw`. Not sure about this yet, but I thought
I'd see how it felt
- I tried to "colocate" env vars so that they're closer to the commands
that use them but I've been going back and forth on it